[RUI-251]: Cleanup files
